refactor: organize routes and components
