+++ src/routes/StatsPage.tsx
@@ -1,8 +1,8 @@
-import { useEffect, useState } from 'react';
-import { getStats, clearStats } from './stats.ts';
-import type { Stats } from './stats.ts';
+import { useEffect, useState, type ReactElement } from 'react';
+import { getStats, clearStats } from '../stats.ts';
+import type { Stats } from '../stats.ts';
 
-function StatsPage() {
+function StatsPage(): ReactElement {
   const [stats, setStats] = useState<Stats>(getStats());
 
   useEffect(() => {
